// src/app/models/case.ts
export interface legal_case {
  id?: number;
  victim: string;
  aggressor: string;
  date: string;
  time: any;
  age: string;
  direction: string;
  latitude: string;
  longitude: string;
  civilStatus: string;
  gender: string;
  font: string;
  temporaryBehavior: string;
  spatialBehavior: string;
  created_at?: string;
  updated_at?: string;
  state?: string;
}